Split up PhysicalCopy into PhysicalCopyFromFile and PhysicalCopyToFile
