+++ src/execution/operator/persistent/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_library_unity(duckdb_operator_persistent OBJECT physical_copy.cpp physical_delete.cpp physical_insert.cpp physical_update.cpp)
+add_library_unity(duckdb_operator_persistent OBJECT buffered_csv_reader.cpp physical_copy_from_file.cpp physical_copy_to_file.cpp physical_delete.cpp physical_insert.cpp physical_update.cpp)
 set(ALL_OBJECT_FILES ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_operator_persistent> PARENT_SCOPE)
