+++ src/execution/physical_plan/plan_copy.cpp
@@ -1,6 +1,7 @@
 #include "catalog/catalog_entry/table_catalog_entry.hpp"
 #include "execution/operator/aggregate/physical_hash_aggregate.hpp"
-#include "execution/operator/persistent/physical_copy.hpp"
+#include "execution/operator/persistent/physical_copy_from_file.hpp"
+#include "execution/operator/persistent/physical_copy_to_file.hpp"
 #include "execution/physical_plan_generator.hpp"
 #include "planner/operator/logical_copy.hpp"
 
@@ -10,9 +11,9 @@ using namespace std;
 unique_ptr<PhysicalOperator> PhysicalPlanGenerator::CreatePlan(LogicalCopy &op) {
 	if (op.children.size() > 0) {
 		auto plan = CreatePlan(*op.children[0]);
-		// COPY from select statement
+		// COPY from select statement to file
 		assert(!op.table);
-		auto copy = make_unique<PhysicalCopy>(op, move(op.info));
+		auto copy = make_unique<PhysicalCopyToFile>(op, move(op.info));
 		copy->names = op.names;
 		copy->sql_types = op.sql_types;
 
@@ -21,6 +22,6 @@ unique_ptr<PhysicalOperator> PhysicalPlanGenerator::CreatePlan(LogicalCopy &op) 
 	} else {
 		// COPY to table
 		dependencies.insert(op.table);
-		return make_unique<PhysicalCopy>(op, op.table, move(op.info));
+		return make_unique<PhysicalCopyFromFile>(op, op.table, move(op.info));
 	}
 }
