+++ src/include/execution/operator/list.hpp
@@ -14,7 +14,8 @@
 #include "execution/operator/join/physical_nested_loop_join.hpp"
 #include "execution/operator/join/physical_piecewise_merge_join.hpp"
 #include "execution/operator/order/physical_order.hpp"
-#include "execution/operator/persistent/physical_copy.hpp"
+#include "execution/operator/persistent/physical_copy_from.hpp"
+#include "execution/operator/persistent/physical_copy_to.hpp"
 #include "execution/operator/persistent/physical_delete.hpp"
 #include "execution/operator/persistent/physical_insert.hpp"
 #include "execution/operator/persistent/physical_update.hpp"
