+++ src/include/execution/operator/persistent/buffered_csv_reader.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// execution/operator/persistent/physical_copy.hpp
+// execution/operator/persistent/buffered_csv_reader.hpp
 //
 //
 //===----------------------------------------------------------------------===//
@@ -12,15 +12,16 @@
 #include "parser/parsed_data/copy_info.hpp"
 
 namespace duckdb {
-class PhysicalCopy;
+class PhysicalCopyFromFile;
 
+//! Buffered CSV reader is a class that reads values from a stream and parses them as a CSV file
 class BufferedCSVReader {
 	static constexpr index_t INITIAL_BUFFER_SIZE = 16384;
-	constexpr static index_t MAXIMUM_CSV_LINE_SIZE = 1048576;
+	static constexpr index_t MAXIMUM_CSV_LINE_SIZE = 1048576;
 public:
-	BufferedCSVReader(PhysicalCopy &copy, std::istream &source);
+	BufferedCSVReader(PhysicalCopyFromFile &copy, std::istream &source);
 
-	PhysicalCopy &copy;
+	PhysicalCopyFromFile &copy;
 	std::istream &source;
 
 	unique_ptr<std::istream> csv_stream;
@@ -52,26 +53,4 @@ private:
 	void Flush(ClientContext &context);
 };
 
-//! Physically copy file into a table
-class PhysicalCopy : public PhysicalOperator {
-public:
-	PhysicalCopy(LogicalOperator &op, TableCatalogEntry *table, unique_ptr<CopyInfo> info)
-	    : PhysicalOperator(PhysicalOperatorType::COPY, op.types), table(table), info(move(info)) {
-	}
-
-	PhysicalCopy(LogicalOperator &op, unique_ptr<CopyInfo> info)
-	    : PhysicalOperator(PhysicalOperatorType::COPY, op.types), table(nullptr), info(move(info)) {
-	}
-
-	void GetChunkInternal(ClientContext &context, DataChunk &chunk, PhysicalOperatorState *state) override;
-
-	//! The table to copy into (only for COPY FROM)
-	TableCatalogEntry *table;
-	//! Settings for the COPY statement
-	unique_ptr<CopyInfo> info;
-	//! The names of the child expression (only for COPY TO)
-	vector<string> names;
-	//! The types of the child expression (only for COPY TO)
-	vector<SQLType> sql_types;
-};
-} // namespace duckdb
+}
